--- install_xspeg.sh
@@ -1,17 +1,0 @@
-#!/bin/bash
-
-
-mkdir ~/bin
-cp ./package/xspeg ~/bin
-
-mkdir ~/local
-mkdir ~/local/xspeg
-cp -r ./package/* ~/local/xspeg
-
-chmod +x ~/local/xspeg/xspeg
-
-echo '#added by XSPEG'>>~/.bash_profile
-echo 'export PATH=$PATH:~/bin/'>>~/.bash_profile
-echo 'export XSPEGLIB=~/local/xspeg'>>~/.bash_profile
-source ~/.bash_profile
-
