--- package/xspeg
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-source ~/.bash_profile
-python $XSPEGLIB/start.py
