--- sample_output/INCAR
@@ -1,20 +1,0 @@
-#parameter input format 
-#first 16 characters: variable name
-#after 17th character: value, in python style
-
-#MASS should be in unit of solar mass
-MASS            1e6
-MASS_RATIO      1e-2
-#SPIN, ECC, P should be in geometrized unit
-SPIN            0.9
-ECC             0.0
-P               10
-IOTA            numpy.pi/4
-METRIC          'Kerr'
-#time should be in unit of second
-TOTAL_TIME      1e4
-
-#observation setting
-THETA           numpy.pi/4
-PHI             0
-R_PC            5e9
