--- sample_output/OUTCAR
@@ -1,23 +1,0 @@
-------------------------------------------
-Reading INCAR file...
-INCAR file read.
-
-Configuration shown below:
-ECC             0.0000000000e+00
-P               1.0000000000e+01
-IOTA            7.8539816340e-01
-SPIN            9.0000000000e-01
-MASS            1.0000000000e+06
-THETA           7.8539816340e-01
-PHI             0.0000000000e+00
-MASS_RATIO      1.0000000000e-02
-R_PC            5.0000000000e+09
-METRIC          KERR
-
-
-------------------------------------------
-Orbit calculation start...
-Orbit Calculation done. Orbits saved in ORBCAR
-------------------------------------------
-Waveform calculation done. Waveform saved in WAVECAR
-------------------------------------------
